Add LayerManagerContext to do layer management and clean up App.jsx
